requirements.txt
customtkinter==5.2.0
pillow==10.0.0
requests==2.31.0
tkcalendar==1.6.1
pywin32==306
pandas==2.1.0 